rename resource name
