+++ libraries/google/dataprocmetastore/property/servicebackup_service_revision_hive_metastore_config_auxiliary_versions.rb
@@ -16,7 +16,7 @@
 module GoogleInSpec
   module DataprocMetastore
     module Property
-      class ProjectLocationServiceBackupServiceRevisionMetadataIntegrationDataplexConfigLakeResources
+      class ServiceBackupServiceRevisionHiveMetastoreConfigAuxiliaryVersions
         attr_reader :additional_properties
 
         def initialize(args = nil, parent_identifier = nil)
@@ -26,7 +26,7 @@ module GoogleInSpec
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationServiceBackupServiceRevisionMetadataIntegrationDataplexConfigLakeResources"
+          "#{@parent_identifier} ServiceBackupServiceRevisionHiveMetastoreConfigAuxiliaryVersions"
         end
       end
     end
