+++ libraries/google/dataprocmetastore/property/servicebackup_service_revision_metadata_integration_data_catalog_config.rb
@@ -16,7 +16,7 @@
 module GoogleInSpec
   module DataprocMetastore
     module Property
-      class ProjectLocationServiceBackupServiceRevisionMetadataIntegrationDataCatalogConfig
+      class ServiceBackupServiceRevisionMetadataIntegrationDataCatalogConfig
         attr_reader :enabled
 
         def initialize(args = nil, parent_identifier = nil)
@@ -26,7 +26,7 @@ module GoogleInSpec
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationServiceBackupServiceRevisionMetadataIntegrationDataCatalogConfig"
+          "#{@parent_identifier} ServiceBackupServiceRevisionMetadataIntegrationDataCatalogConfig"
         end
       end
     end
