+++ libraries/google/dataprocmetastore/property/servicebackup_service_revision_metadata_integration_dataplex_config_lake_resources.rb
@@ -16,7 +16,7 @@
 module GoogleInSpec
   module DataprocMetastore
     module Property
-      class ProjectLocationServiceBackupServiceRevisionHiveMetastoreConfigConfigOverrides
+      class ServiceBackupServiceRevisionMetadataIntegrationDataplexConfigLakeResources
         attr_reader :additional_properties
 
         def initialize(args = nil, parent_identifier = nil)
@@ -26,7 +26,7 @@ module GoogleInSpec
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationServiceBackupServiceRevisionHiveMetastoreConfigConfigOverrides"
+          "#{@parent_identifier} ServiceBackupServiceRevisionMetadataIntegrationDataplexConfigLakeResources"
         end
       end
     end
