+++ libraries/google/dataprocmetastore/property/servicebackup_service_revision_scheduled_backup.rb
@@ -13,24 +13,36 @@
 #     CONTRIBUTING.md located at the root of this package.
 #
 # ----------------------------------------------------------------------------
-require 'google/dataprocmetastore/property/projectlocationservicebackup_service_revision_network_config_consumers'
+require 'google/dataprocmetastore/property/servicebackup_service_revision_scheduled_backup_latest_backup'
 module GoogleInSpec
   module DataprocMetastore
     module Property
-      class ProjectLocationServiceBackupServiceRevisionNetworkConfig
-        attr_reader :consumers
+      class ServiceBackupServiceRevisionScheduledBackup
+        attr_reader :enabled
 
-        attr_reader :custom_routes_enabled
+        attr_reader :cron_schedule
+
+        attr_reader :time_zone
+
+        attr_reader :next_scheduled_time
+
+        attr_reader :backup_location
+
+        attr_reader :latest_backup
 
         def initialize(args = nil, parent_identifier = nil)
           return if args.nil?
           @parent_identifier = parent_identifier
-          @consumers = GoogleInSpec::DataprocMetastore::Property::ProjectLocationServiceBackupServiceRevisionNetworkConfigConsumersArray.parse(args['consumers'], to_s)
-          @custom_routes_enabled = args['customRoutesEnabled']
+          @enabled = args['enabled']
+          @cron_schedule = args['cronSchedule']
+          @time_zone = args['timeZone']
+          @next_scheduled_time = args['nextScheduledTime']
+          @backup_location = args['backupLocation']
+          @latest_backup = GoogleInSpec::DataprocMetastore::Property::ServiceBackupServiceRevisionScheduledBackupLatestBackup.new(args['latestBackup'], to_s)
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationServiceBackupServiceRevisionNetworkConfig"
+          "#{@parent_identifier} ServiceBackupServiceRevisionScheduledBackup"
         end
       end
     end
