--- test/integration/verify/controls/google_dataproc_metastore_project_location_service_backup.rb
@@ -1,43 +1,0 @@
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in README.md and
-#     CONTRIBUTING.md located at the root of this package.
-#
-# ----------------------------------------------------------------------------
-
-title 'Test GCP google_dataproc_metastore_project_location_service_backup resource.'
-
-gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
-project_location_service_backup = input('project_location_service_backup', value: {
-  "name": "value_name",
-  "parent": "value_parent",
-  "create_time": "value_createtime",
-  "end_time": "value_endtime",
-  "state": "value_state",
-  "description": "value_description"
-}, description: 'project_location_service_backup description')
-control 'google_dataproc_metastore_project_location_service_backup-1.0' do
-  impact 1.0
-  title 'google_dataproc_metastore_project_location_service_backup resource test'
-
-  describe google_dataproc_metastore_project_location_service_backup(name: project_location_service_backup['name']) do
-  	it { should exist }
-  	its('name') { should cmp project_location_service_backup['name'] }
-  	its('create_time') { should cmp project_location_service_backup['create_time'] }
-  	its('end_time') { should cmp project_location_service_backup['end_time'] }
-  	its('state') { should cmp project_location_service_backup['state'] }
-  	its('description') { should cmp project_location_service_backup['description'] }
-
-  end
-
-  describe google_dataproc_metastore_project_location_service_backup(name: "does_not_exit") do
-  	it { should_not exist }
-  end
-end
